refactor: change event-bus filename to EventBus
