+++ src/renderer/components/Audio.vue
@@ -19,7 +19,7 @@
 </template>
 
 <script>
-import { EventBus } from '@/utils/event-bus'
+import { EventBus } from '@/utils/EventBus'
 
 export default {
   data() {
