+++ src/renderer/components/drawer/Drawer-timer.vue
@@ -95,7 +95,7 @@
 </template>
 
 <script>
-import { EventBus } from '@/utils/event-bus'
+import { EventBus } from '@/utils/EventBus'
 
 export default {
   name: 'Drawer-timer',
