+++ src/renderer/components/timer/Timer-controller.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import { EventBus } from '@/utils/event-bus'
+import { EventBus } from '@/utils/EventBus'
 
 export default {
   computed: {
