+++ src/renderer/components/timer/Timer-footer.vue
@@ -122,7 +122,7 @@
 </template>
 
 <script>
-import { EventBus } from '@/utils/event-bus'
+import { EventBus } from '@/utils/EventBus'
 
 export default {
   name: 'TimerFooter',
