+++ src/renderer/components/timer/Timer.vue
@@ -120,7 +120,7 @@ import appTrayIcon from '@/components/TrayIcon'
 import appTimerController from '@/components/timer/Timer-controller'
 import appTimerDial from '@/components/timer/Timer-dial'
 import appTimerFooter from '@/components/timer/Timer-footer'
-import { EventBus } from '@/utils/event-bus'
+import { EventBus } from '@/utils/EventBus'
 
 export default {
   components: {
